requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.13
django-allauth==0.63.3
django-cleanup==8.1.0
gunicorn==20.1.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
whitenoise==5.3.0
requests
cryptography
django-summernote-0.8.20.0